Allow ReplicaSet and Job metadata generators to use partial meta objects (#109)
We only use metadata from Jobs and ReplicaSets, but require that full
resources are supplied. This change relaxes this requirement, allowing
PartialObjectMetadata resources to be used. This allows callers to use
metadata informers and avoid having to receive and deserialize
non-metadata updates from the API Server.

See https://github.com/elastic/elastic-agent/pull/5580 for an example of
how this could be used. I'm planning to add the metadata informer from
that PR to this library as well. Together, these will allow us to
greatly reduce memory used for processing and storing ReplicaSets and
Jobs in beats and elastic-agent.

This is will help https://github.com/elastic/elastic-agent/pull/5580 and
https://github.com/elastic/elastic-agent/issues/4729 specifically, and
https://github.com/elastic/elastic-agent/issues/3801 in general.