Add metadata watcher and informer (#111)
Introduce metadata-only watchers to the kubernetes package. These are
useful if we only need to track metadata for a resource - a good example
are ReplicaSets, for which we usually only care about the
OwnerReferences. As a result, we only store the metadata, reducing
steady-state memory consumption, but also only get updates involving
metadata, reducing churn greatly in larger clusters.

The implementation introduces new constructors for the Watcher, allowing
an informer to be passed in. Existing constructors are implemented using
the new constructor, though none of the code actually changes. As a
result, it is now possible to unit test the watcher, and I've added some
basic unit tests for it.

We also add two helper functions:

- `GetKubernetesMetadataClient` creates a metadata-only kubernetes
client, and is very similar to the existing `GetKubernetesClient`
- `RemoveUnnecessaryReplicaSetData` is a transform function that can be
passed into an informer so it only stores the metadata we actually use

I tested these new functions in both beats and agent, in a kind cluster
as well as one of our staging clusters.

This is part of the solution to
https://github.com/elastic/elastic-agent/pull/5580.

---------

Co-authored-by: Mauri de Souza Meneguzzo <mauri870@gmail.com>